requirements.txt
fastapi
docker
python-dotenv
uvicorn